web_interface/backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Cors==4.0.0
scikit-learn==1.3.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
Pillow==10.0.0